Modify the xorpsh message when the configuration is changed by some other user: "had been changed" -> "has been changed".
If we are to be pedantic, as it was explained to me, the original
form "had been" is correct if we consider the fact that the
configuration was changed in the past (probably <1 second ago).
If we ignore the communication latency, then the new "has been"
form is preferred (to emphasize the fact that the change just
happened).

This fixes Bugzilla entry #595.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=595
Submitted by:	Kristian Larsson
Approved by:	Atanu Ghosh


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8464 26435e26-4b25-483a-942d-477236992e84