Port a Windows-specific fix in fea/rawsock.cc back to the MFEA:
When opening a raw socket for receiving IPv4 datagrams on
Windows, bind it to INADDR_ANY before doing anything else.
Remove the code which previously did this only whilst
joining an IPv4 multicast group.

This resolves a critical bug in the Windows FEA, whereby if a
raw socket was opened, but not joined to a multicast group, the
receive path internal to Winsock would not be hooked up and thus
no unicast or multicast traffic could be received.

[Note that currently the MFEA doesn't work on Windows, but we want
to preserve both rawsock implementations consistent until the
MFEA starts using the FEA's rawsock.]

Submitted by:	Bruce M. Simpson


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8478 26435e26-4b25-483a-942d-477236992e84