Remove any libcomm dependencies, because nothing inside the libxorp directory must depend on any other XORP library.
Unfortunately, as a result of this we have to copy few of the
libcomm functions inside test_asyncio.cc.
Currently, those functions are preserved almost unmodified
in case we need to perform any maintainance on them.
In the future they could be trimmed-down for the particular purpose
they are used inside test_asyncio.cc.

Bug found by:	Atanu Ghosh


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8481 26435e26-4b25-483a-942d-477236992e84