Modify the following XRLs to include all known information about deleted routes:
redist4/0.1/delete_route
redist6/0.1/delete_route
redist_transaction4/0.1/delete_route
redist_transaction6/0.1/delete_route

Now the arguments for the above delete_route XRLs are exactly same
as the corresponding add_route XRLs.
The additional information is: nexthop, ifname, vifname, metric,
and admin_distance.

Some of this information is needed by Windows when deleting the
routes in the kernel: nexthop, ifname/vifname.
The metric and admin_distance information is not really needed for now,
but is included for consistency, and just in case if in the future we
need it for some special purpose.

Requested by:	Bruce M. Simpson
Approved by:	Atanu Ghosh


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8483 26435e26-4b25-483a-942d-477236992e84