Remove the mechanism for having two (Linux-specific) netlink sockets NetlinkSocket4 and NetlinkSocket6 for IPv4 and IPv6-related communications with the kernel, and use only the (new) neutral NetlinkSocket class instead.
The reason there were NetlinkSocket4 and NetlinkSocket6 is
the following. According to the netlink(7) manual page, netlink
family NETLINK_ROUTE is suppose to be used to receive and
send IPv4 routing table updates, while NETLINK_ROUTE6 is suppose
to be used for IPv6 routing table updates.

The manual page appears to be wrong: only NETLINK_ROUTE should
be used (for both IPv4 and IPv6). This was also discovered by trial
and error when the NetlinkSocket[4,6] code was developed and tested,
and in fact NETLINK_ROUTE6 was never used.

This eliminates the need for two classes NetlinkSocket4 and
NetlinkSocket6, which simplifies significantly some of the
netlink-related the code.

[For the record, the netlink(7) manual page has been corrected by
Hasso Tepper: see Bugzilla entry #463]

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=464
Submitted by:	Hasso Tepper


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8484 26435e26-4b25-483a-942d-477236992e84