Fix a problem with withdrawing routes from the Windows FIB. The problem is only exposed when multiple routing protocols are run which exchange exactly the same route prefixes.
Make sure that the FEA bottom half completely fills out
the MIB_IPFORWARDROW structure before passing it to
DeleteIpForwardEntry(), on Windows Server.

This change relies on the recent changes to the redist
XRL interface to pass interface name and next-hop information
to the FEA when withdrawing routes.

While we're here, reintroduce the debug_msg() which instruments
this path, and kill the old code which relied on GetBestRoute()
for constructing the MIB_IPFORWARDROW structure.

Significantly update code comments.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8487 26435e26-4b25-483a-942d-477236992e84