Explicitly check for python in the environment, and use it if found. If not found, die.
This is a defensive measure to ensure that python is installed
when running the BGP regression tests.

Currently we are not passed $PYTHON from the top level make,
although we use AC_CHECK_PROG for Python. We should really
use AM_PATH_PYTHON, and propagate the $PYTHON variable from
GNU Make into the environment, to correctly use the detected
python binary from the build environment.

For cross-compiles, this would be a no-op.

Tested on FreeBSD 6.1 and Windows Server 2003.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8533 26435e26-4b25-483a-942d-477236992e84