Fix an edge case in the BGP regression tests.
The test_peering2.sh set of regression tests depends upon the
coord.exe and peer.exe processes in its implementation.

Due to the nature of how they are invoked from the MSYS bash shell,
it is necessary to perform expansion of the '/tmp' path, as well
as UNIX to NT path conversion, from within the executable itself.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8542 26435e26-4b25-483a-942d-477236992e84