Remove the sort_by_value() method and always use the sort_by_template() mechanism, because sort_by_value() is not really useful.
The only difference between them is that the former leaves
nodes unsorted if they have different templates.
The sort_by_value() method has been (mis)used in several places
and the result of this is undeterministic ordering of some of the
configuration nodes (e.g., leaf nodes with values may be in the
order they have been added by xorpsh).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8554 26435e26-4b25-483a-942d-477236992e84