Increase the default value for table-request-secs from 1 to 30 (seconds). This parameter is used to specify the frequency of transmitting the Request messages on an interface if there are no other RIP/RIPng neighbors.
The original frequency of 1 (second) is an overkill, and 30 seconds
frequency should be sufficient, especially given that the first Request
message is transmitted right after startup.

Note that this frequency (and even the fact that the Request messages
must be periodic when there are no neighbors) are not in the RIP/RIPng
spec (RFC 2453 and RFC 2080 respectively).

This fixes Bugzilla entry #613.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=613
Submitted by:	Justin Fletcher <jfletcher AT vyatta.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8629 26435e26-4b25-483a-942d-477236992e84