Expand the test_execute_terminate_command() test for execution and termination of a command. Now we test whether the termination succeeds.
Previously, there was a run-time race condition observed
on Solaris-10.0 (i386) where appeared that the termination signal
was delivered not only to the current command but to the command
from the next test as well. E.g., the race condition would
go away if there was just a very short sleep right after the command
termination inside test_execute_terminate_command().


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8651 26435e26-4b25-483a-942d-477236992e84