Fix Origin and MED attribute processing in BGP route aggregation.
When aggregating BGP routes the aggregate will contain an AS SET containing
all non-matching AS numbers (somewhat oversimplified algorithm description),
instead of simply originating an empty AsPath with the ATOMIC AGGREGATE
attribute set, as was done previously.

Given that the "brief" aka ATOMIC AGGREGATE mode is much simpler and
works faster, implementing an interface which would allow an
administrator to choose which aggregation method to use is now on my
TODO list, after the UNH tests are done.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8670 26435e26-4b25-483a-942d-477236992e84