Reverse-back the previous commit. The original code's semantics was fine, because the writing/reading of variables was happening by utilizing a local policy cache. The reading was actually returning the new value as written to the cache by an earlier policy term.
The sync() was necessary at the end of a policy statement to write
the final values to the underlying structure.

In fact, the previous commit broke the (desired) syntax, because
each sync() will throw-away the cache, so the read (from the
empty cache) wouldn't return the expected value.

Requested by:	Andrea Bittau


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8715 26435e26-4b25-483a-942d-477236992e84