Bugfix: RedistXrlOutput<A>::task_completed() was trying to start dequeuing XRL requests even if the request queue was empty, which was triggering an assert at the beginning of RedistXrlOutput<A>::start_next_task()
It seems that it is possible for the XRL router to occassionaly fail
to accept a new request even if there are no other XRLs in flight.
This was observed through an assert in the
RedistXrlOutput<A>::start_next_task() dequeuing loop being triggered
several times on pal.  As a temporary measure this assert is now
converted to a warning: we need to understand better when and why
exactly the dispatching of XRL requests can occasionaly fail.

Moreover, if the dispatch fails with no requests in flight (and thus
no callbacks pending which could later resume the dequeuing), a
synthetic pause is generated which will attempt to retry dispatching
the original request after 10 ms.

Bug found by:	Atanu


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8787 26435e26-4b25-483a-942d-477236992e84