If the ConfigTreeNode operator is invalid, then immediately return an error rather than calling show_operator() and operator_to_str() which will coredump with XLOG_UNREACHABLE().
Note that this error won't be triggered in xorpsh, because xorpsh
prohibits setting and committing invalid values to config nodes.
It might happen in programs that use some of the xorpsh code as
a library and modify the configuration in slightly different way.
E.g., the Vyatta GUI xgdaemon allows setting various values to
the configuration, but right before committing it calls
ConfigTreeNode::check_config_tree() to verify the values (which itself
will trigger the coredump).

Submitted by:	Marat Nepomnyashy <marat AT vyatta.com>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8868 26435e26-4b25-483a-942d-477236992e84