If an interface's "enable" status is flipped internally (i.e., the interface was enabled, then it was disabled and enabled again) for the purpose of writing some other configuration to the kernel (e.g., the MAC address or the MTU), then report this change to all parties that are listening for interface status (e.g., libfeaclient).
This fixes a bug when modifying the MAC address of an interface
removes XORP routes from the kernel, because on some OS (e.g,, Linux)
disabling an interface automatically removes the kernel routes
that use that interface.

Previously, the routing protocols didn't realize that their routes
were automatically removed from the kernel because of the internal
interface flipping. By propagating the disable/enable to the routing
protocols, they will learn about the change.
Then, eventually each routing protocol will try to delete its routes
that use that interface when it learns that the interface is disabled.
After it learns that the interface is again enabled, the routing
protocol will try to introduce again those routes.

Bug found by:	Vyatta


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@8945 26435e26-4b25-483a-942d-477236992e84