Decrease BACKSPACE_DEPTH from 4 to 2.
If BACKTRACE_DEPTH is larger than 2, then the program coredumps
on amd64 + FreeBSD-6.1.
If it is larger than 4, the program coredumps on i386 + FreeBSD-4.10.

Note that the mechanism for getting the backspace implemented
inside get_backtrace() is not robust enough, hence we need this
hand-crafting of BACKSPACE_DEPTH.

See the following discussion about the issue itself:
http://www.codecomments.com/archive286-2004-12-348859.html
Also, see the glibc implementation of backtrace(3)
(e.g., inside "debug/backtrace.") for some ideas how to implement it.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9036 26435e26-4b25-483a-942d-477236992e84