Reverse the previous commit. When setting the MAC address using the RTM_SETLINK netlink mechanism, the attribute payload indeed is suppose to be a MAC address (only).
The Linux kernel (at least version 2.6.17.10) indeed had a bug which
required the MAC address to be encapsulated inside "struct sockaddr".
See the following email thread for details:
http://www.spinics.net/lists/kernel/msg498873.html

The Linux kernel has been fixed, and the fix will be available
with kernel 2.6.18 (if not earlier).

In the mean time, the simplest work-around to have the FEA working
with kernels that aren't fixed is to replace "#ifdef RTM_SETLINK"
with "#if 0" inside IfConfigSetNetlink::set_interface_mac_address().


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9155 26435e26-4b25-483a-942d-477236992e84