Add a new template class AlignData for aligning buffer data with a particular data type.
The technically correct solution is to allocate (using malloc())
new buffer and copy the original data to it. By definition, the
malloc()-ed data is aligned, and therefore it can be casted to the
desired type.

The more efficient solution (but probably technically incorrect),
is to assume that the first byte of "vector<uint8_t>" buffer is aligned
similar to malloc()-ed data, and therefore it can be casted to the
desired type without creating a copy of it.

The desired behavior can be chosen by setting the
AlignData::_data_is_copied constant to true or false.
Note that the constant is predefined for all AlignData instances.
If necessary, the constant can become a variable that can have
different value for each AlignData instance.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9180 26435e26-4b25-483a-942d-477236992e84