The network layer reachability information in a BGP update packet is coded as one or more 2-tuples, the first is the length in bits and the second is the IP prefix itself. The default route 0.0.0.0/0 has a zero in the length part of the tuple and no prefix.
The BGPUpdateAttrib class which is responsible for parsing an NLRI
was unconditionally reading the prefix byte and shifting it left 24
bits. BGP receiving a default route could end up with an NLRI of the
form 42.0.0.0/0, this is obviously illegal and should not be possible,
however, the BGPUpdateAttrib class inherits from IPv4Net and directly
writes to its member variables bypassing any sanity checking. The
member variable of IPv4Net should be made private to prevent this
problem in the future, see <http://www.xorp.org/bugzilla/show_bug.cgi?id=654>.

This problem was first seen when BGP received a default route and an
assert was hit in the ref trie code due to the illegal nature of the
prefix and prefix length.

A test function test_announce_packet2 has also been added.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9227 26435e26-4b25-483a-942d-477236992e84