Reverse the previous work-around for DragonFlyBSD-1.6 where we conditionally don't include <netinet/if_ether.h>.
The problem with DragonFlyBSD-1.6 file /usr/include/machine/atomic.h
is also triggered when we include <net/if_var.h>, and this file is
included in number of places by the FEA, so there is no clean work-around
for that problem.

Instead, the users should patch the problematic
/usr/include/machine/atomic.h as specified in the BUILD_NOTES by
applying the following patch:

http://www.dragonflybsd.org/cvsweb/src/sys/i386/include/atomic.h.diff?r1=1.20.2.1&r2=1.20.2.2

Note that the problem has been fixed in the DragonFlyBSD source tree
so eventually it will appear in the next release.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9243 26435e26-4b25-483a-942d-477236992e84