Fix a bug related to managing the IP addresses for the multicast related modules.
When updating the set of interfaces/vifs and their addresses,
first add/update new and existing entries (interfaces/vifs/addresses),
and after that delete obsolete entries.

This fixes an issue when atomically changing the IP address of a router
would set an interface DOWN if the (single) old address is deleted
before adding the new address.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9258 26435e26-4b25-483a-942d-477236992e84