Fix the BGP Peer OIDs.
It appears that Net-SNMP has a (semantic) bug when using ASN_IPADDRESS
values as an index in a table.
E.g., ASN_IPADDRESS values in general are suppose to be stored in
network order, but it appears they have to be stored in host order
if are to be used as an index in a table.

See the following URL for a related discussion (within UDP context);
http://groups.google.com/group/mailing.unix.net-snmp-coders/browse_thread/thread/24bc2c2aed8663bc/e8837358d13f0de6

This fixes Bugzilla entry #662.

Bugzilla URL:	http://www.xorp.org/bugzilla/show_bug.cgi?id=662


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9285 26435e26-4b25-483a-942d-477236992e84