If there is an error when deleting a forwarding entry, then check the reason for the error and eventually ignore it.
For example, if the outgoing interface was taken down earlier, then
most likely the kernel has removed the matching forwarding
entries on its own. Hence, check whether all of the following
is true:
  - the error code matches
  - the outgoing interface is down

If all conditions are true, then ignore the error and consider
the deletion was success.
Note that we could add to the following list the check whether
the forwarding entry is not in the kernel, but this is probably
an overkill.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9336 26435e26-4b25-483a-942d-477236992e84