Fix a coredump problem during the xorp_rtrmgr shutdown if a statement list like network4-list is used.
The issue was that a Dependancy instance (Configuration::_sets) was
used (by a reference) after it has been destroyed.

The solution is to clear the Configuration::_policies before the
Configuration::_sets goes out of scope.
Otherwise, the _policies destructor might try to use a reference
to _sets after _sets has been destroyed.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9354 26435e26-4b25-483a-942d-477236992e84