Add a CacheTable after the PolicyTableSourceMatch policy table (before the AggregationTable).
This modification is needed as an intermediate solution to solve
the following problem.

If a new peer comes up such that it is handled by DumpTable,
and if a route is modified by the PolicyTableSourceMatch table, then
an InternalMessage that is "changed" by the PolicyTableSourceMatch table
might reach the BGPRouteFilter table. This table doesn't expect to see
any InternalMessage that is "changed" (there are a number of asserts
to prevent that), so BGP would coredump.
Simply removing the asserts is not sufficient, because in some cases the
BGPRouteFilter table memory management is not consistent with the
PolicyTableSourceMatch memory management of the route.
E.g., BGPRouteFilter<A>::drop_message() will unref the original route
if it came modified.

Adding the CacheTable after the PolicyTableSourceMatch policy table
is used to smooth-out the "changed" InternalMessage(s) so the
BGPRouteFilter table never sees an InternalMessage that is changed.

All the BGP regression tests appear to pass with this modification.

Approved by:	Atanu Ghosh


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9452 26435e26-4b25-483a-942d-477236992e84