If we stop running PIM on an interface, then don't attempt to inform other PIM neighbors that this interface is not more the Designated Router by sending a PIM Hello message with DR Priority option of 0.
This optimization is not in the spec and is not necessary in most cases,
because the final PIM Hello message with Holdtime of 0 (as mandated by
the PIM-SM spec) will have the same effect.

Only if there are lots of joined multicast groups that might take some
time for the PIM-SM router to prune before shutting down, then it might
be useful to tell immediately everyone to elect some other DR.
However, doing this properly adds complexity and might create
confusion (again, because it is not part of the PIM-SM spec).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9530 26435e26-4b25-483a-942d-477236992e84