In OSPFv3 the Router-LSAs and Network-LSAs no longer carry any addressing information, so they next hop information must be extracted from Link-LSAs or from the source address of a neighbour hello packet.
Add support for finding the source address of a neighbour.

The set IPv4 address method in the vertex previously took a uint32_t
and converted it to an IPv4, change this to take an IPv4
directly. This change was made as the first change required the
unnecessary specialisation of a number of methods.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9595 26435e26-4b25-483a-942d-477236992e84