Fix a bug when adding an IPv6 unicast forwarding entry to the kernel on KAME-derived systems if this is an interface-specific route and the next-hop is a link-local address.
In case of KAME sometimes the local interface index is encoded
in the third and fourth octet of an IPv6 address (for link-local
unicast/multicast addresses) when we add an unicast route to the kernel.
E.g., see /usr/src/sbin/route/route.c on FreeBSD-6.2 and search for
the __KAME__ marker.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9729 26435e26-4b25-483a-942d-477236992e84