Support for advertising multiple interface/vif global addresses.
It is no longer necessary to configure an interface address, if no
addresses are configured all the global addresses that belong to the
interace/vif will be advertised.

If one or more addresses are configured then only the configured
addresses will be advertised.

To stop any global addresses being advertised set an interface address
and disable it.

Note this change breaks OSPFv3 boot files as interface/vif parameters
are no longer in the address node, there was a warning in the release
notes.

Configurations that did not set priority, hello-interval,
router-dead-interval, interface-cost, retransmit-interval,
transit-delay or passive will continue to work.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9737 26435e26-4b25-483a-942d-477236992e84