Monitor the change of status of an interface/vif previously only the status change of an interface/vif/address was being monitored.
In OSPFv2 as only one address is allowed per interface/vif an
interface/vif/address being disabled meant that the interface had gone
down. In OSPFv3 an address being removed from an interface can send
this signal, so removing an address from an interface was taking the
interface down.

Now an interface/vif going down is a signal that the interface has
gone down.

In OSPFv2 an interface/vif/address going down is ignored and for
OSPFv3 is a signal to recompute the router Link-LSA.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9820 26435e26-4b25-483a-942d-477236992e84