Bring in configurable administrative distance feature to the RIB.
As discussed with Atanu and Pavlin this is only the first phase of
implementing administrative distance. The RIB does not yet recompute
routes when admin distance is changed. This is good enough to get XORP
into a state where we can start to support protocol modules from
external contributors.

Pending regen of am-files on HEAD.

Bugzilla ID:	536


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9827 26435e26-4b25-483a-942d-477236992e84