* Change the priority of the "Pull Routes Task" from PRIORITY_HIGH   to PRIORITY_DEFAULT. Otherwise, handling this task might keep   BGP busy for very long time (e.g., 40 seconds or more) without   processing socket events (including XRLs).   If we don't handle promptly the XRLs then the XRL Finder   will declare BGP being dead.
* Decrease the async reader priority of the TCP sockets that handle
  the BGP peering to PRIORITY_LOWEST so it is lower than the
  "Pull Routes Task" priority. Otherwise, on startup the
  handling of high volume data from the BGP peers might delay
  the "Pull Routes Task" handling for a number of minutes.
  The result of that delay would be that no BGP routes will
  be send to the RIB (and the kernel).

  Also, decrease the corresponding async writer priority from
  PRIORITY_HIGH to PRIORITY_DEFAULT, because it just has to be
  better priority than the async reader.

Note that the above refactoring of the priorities is far from
being ideal and it might create some other issues (e.g., delaying the
reading of BGP data for too long while handling processing of
various internal events).
Further refactoring/adjusting of the priorities needs to be done by
someone who is familiar with the details of the BGP implementation.
Alternatively, the eventloop needs to be refactored such that the
processing of one type of events does not starve other types.

This fixes Bugzilla entry #704.

Bug found by:	Vyatta


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9869 26435e26-4b25-483a-942d-477236992e84