Improve the logic for handling handling multicast vifs if PIM receives notifications about network interface related changes:
* If the vif's primary address or the domain-wide address is deleted
  or there are no more addresses then stop the vif.

* If the vif's primary or domain-wide address has changed then
  restart the vif.

Note that the above changes might not work well in practice
in case the kernel decides to do something on its own with
the correponding multicast vif state inside the kernel itself.
Another potential issue is if the kernel doesn't let the userland
program to (re)join the same (IPv4) multicast group with the new
IPv4 address as a handler (probably because of a kernel bug
if the kernel still keeps multicast membership information
about the old IP address inside the network interface driver).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9924 26435e26-4b25-483a-942d-477236992e84