Remove the following XRLs, because they are not needed: - ifmgr/0.1/register_client - ifmgr/0.1/unregister_client - ifmgr/0.1/register_system_interfaces_client - ifmgr/0.1/unregister_system_interfaces_client
The libfeaclient mechanism should be used for monitoring the
status of network interfaces.
Note that currently the libfeaclient mechanism monitors only the
XORP-configured network interfaces, but does not monitor all
network interfaces. If there is the need to monitor all network
interfaces, then the libfeaclient mechanism can easily be extended
to provide that functionality.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@9927 26435e26-4b25-483a-942d-477236992e84