Added a new boolean flag "ip_internet_control" when sending/receiving IP raw packets.
When transmitting IP raw packets, if the flag is true, then this
is IP control traffic and the IPv4 header's ip_tos will be set to
IPTOS_PREC_INTERNETCONTROL.

When receiving IP raw packets, if the IPv4 ip_tos is set to
IPTOS_PREC_INTERNETCONTROL, then the ip_internet_control flag is
set to true.

For now, the flag is ignored for IPv6.

Previously, the is_router_alert (or ip_router_alert) flag was
overloaded to set the ip_tos to IPTOS_PREC_INTERNETCONTROL.
However, this was an issue for some protocols that don't use the
IP Router Alert flag, because they still need to set ip_tos
to IPTOS_PREC_INTERNETCONTROL.

Note that at the bottom of the FEA the ip_tos IP header field
is set according to ip_internet_control ONLY if the ip_tos
XRL argument is not used (i.e., when it has negative value).

Bug found by:	Hasso Tepper <hasso AT estpak.ee>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10019 26435e26-4b25-483a-942d-477236992e84