FEA/MFEA refactoring: get rid of the raw sockets I/O inside the MFEA, and use the FEA instead.
Modify accordingly the MLD/IGMP and PIM-SM modules to use the FEA
instead for raw I/O.

Note that after the modification only PIM-SM needs to register
with the MFEA (per interface) for the purpose of multicast routing.
Also, unlike before there could be only one module/target that
can register (as an owner) per interface with the MFEA.
This is also a step toward implementing a module with the
interoperability rules for multicast routing protocols
(as per RFC 2715).

Note that the refactored code doesn't contain anymore the hack
for sharing the IGMP raw socket and the (MRT_INIT) multicast routing
socket. That hack was needed because Linux seems to use
the MRT_INIT multicast routing socket (only) to pass IGMP messages
to userland (if the IGMP messages don't contain the Router Alert option
and the host hasn't joined the multicast group for the IGMP destination
address).
This issue needs further investigation and eventually will be fixed
independently.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10039 26435e26-4b25-483a-942d-477236992e84