Refactor the mechanism for receiving the multicast forwarding related upcalls from the system.
Previously the MFEA opened its own IGMP or ICMPv6 socket, and
used it for the sole purpose of THE MRT_INIT multicast forwarding
socket (in UNIX there can be only one multicast forwarding socket
per family).
At the same time there was another IGMP or ICMPv6 socket for the purpose
of sending/receiving control packets by the IGMP/MLD userland protocol.

While this was the right design because the two functionalities are
separated, it created a problem on Linux which doesn't propagate
some of the IGMP control packets on the regular IGMP raw socket.
E.g., IGMP Join messages without the Router Alert option and that
are sent to the joined multicast group (which typically the router
is not member of) are propagated by the Linux kernel only to the
MRT_INIT multicast forwarding socket.

To get around this problem, now the MFEA registers for multicast
forwarding related upcalls with the IoIpManager.
The IoIpManager now understands such upcalls, and it will open
only a single IGMP socket that the MFEA itself will change
to MRT_INIT multicast forwarding socket.
Then, the IoIpManager will demultiplex the upcalls to the MFEA,
and the real IGMP messages to the registered external receivers
(e.g., the IGMP module).


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10064 26435e26-4b25-483a-942d-477236992e84