Bug fix in the sysctl(3) mechanism usage.
The size of the obtained data wasn't calculated properly when we
call sysctl(3) twice one after the other (the first time to calculate
the necessary buffer size to store the data).
The correct data size should be obtained from the second time sysctl(3)
is called.

The bug was revealed on NetBSD-3.1 and OpenBSD-4.1.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10089 26435e26-4b25-483a-942d-477236992e84