Continuing with the FEA refactoring.
Now there are data plane managers (one per system) which control
the plugins at the bottom of the FEA.
The data plane managers allocate (or load) the plugins and
register them with the generic IfConfig and FibConfig controllers.

In addition to managing the plugins, the data plane managers are the
entry point for plugin-specific configuration. E.g., all the
Click-specific configuration is passed to the Click data plane manager,
so the rest of the system doesn't know anything about Click.

Note that the design and implementation was done in collaboration
with Adam Greenhalgh <a.greenhalgh AT cs.ucl.ac.uk>

Approved by:	Adam Greenhalgh <a.greenhalgh AT cs.ucl.ac.uk>


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10136 26435e26-4b25-483a-942d-477236992e84