Bug fix with the previous commit.
In the previous commit there was replacement of:
OLD:
  my_map[key] = value;
NEW:
  my_map.insert(make_pair(key, value));

However, those operations are NOT equivalent.
In the second operation the insert will fail if there was already
an entry with same key (i.e., the value is not overwritten).
In the first operation the value will always be installed, eventually
overwritting any old values.

After the fix, we call my_map.erase(key) before calling
my_map.insert(...)

Note that we avoid calling the obvious "my_map[key] = value", because
in this particular case "value" has a little bit odd type:
it is an STL iterator, and MacOS X 10.4.10 with gcc-4.0.0
compiled with -O2 doesn't like it.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10179 26435e26-4b25-483a-942d-477236992e84