Added a check whether the build environment has netlink socket RTA_TABLE message attribute.
This attribute is needed if the routing/forwarding table ID (inside
the Linux kernel) is larger than 0xff.
For the record, the semantic for using the table ID is:

If the table ID is <= 0xff, then we set it in rtmsg->rtm_table,
otherwise we set rtmsg->rtm_table to RT_TABLE_UNSPEC and add the
real value as an RTA_TABLE attribute.

Note that table ID support for values large than 0xff has been
added to Linux around kernel version 2.6.20 or so.


git-svn-id: https://xorp.svn.sourceforge.net/svnroot/xorp@10190 26435e26-4b25-483a-942d-477236992e84